Move format_vat to allow translations, add translations
