+++ src/onegov/org/layout.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+import numbers
+
 import babel.dates
 import re
 
@@ -639,6 +641,29 @@ class Layout(ChameleonLayout, OpenGraphMixin):
 
     def format_seconds(self, seconds: float) -> str:
         return self.format_timedelta(timedelta(seconds=seconds))
+
+    def format_vat(self, amount: numbers.Number | Decimal | float | None,
+                   currency: str = 'CHF') -> str:
+        """
+        Takes the given amount and currency returning the VAT string if the
+        VAT rate is set in the organization settings. The VAT string can be
+        placed right after a price value.
+        """
+        vat_rate = Decimal(self.app.org.vat_rate or 0.0)
+
+        if amount is not None and vat_rate:
+            if isinstance(amount, (Decimal, int, float, str)):
+                amount = Decimal(amount)
+            else:
+                amount = Decimal(str(amount))
+
+            vat = amount * vat_rate / 100
+            vat_name = self.request.translate(_('VAT'))
+            vat_str = (f'({vat_name} {self.format_number(vat_rate, 1)}%'
+                       f' enthalten: {self.format_number(vat)} {currency})')
+            return vat_str
+
+        return ''
 
     def password_reset_url(self, user: User | None) -> str | None:
         if not user:
