+++ src/onegov/org/views/settings.py
@@ -533,7 +533,7 @@ def handle_ticket_data_deletion_settings(
 
 @OrgApp.form(
     model=Organisation, name='vat-settings', template='form.pt',
-    permission=Secret, form=VATSettingsForm, setting=_('VAT'),
+    permission=Secret, form=VATSettingsForm, setting=_('Value Added Tax'),
     icon='fa-file-invoice-dollar', order=450)
 def handle_vat_settings(
         self: Organisation,
@@ -541,8 +541,9 @@ def handle_vat_settings(
         form: VATSettingsForm,
         layout: SettingsLayout | None = None
 ) -> RenderData | Response:
-    layout = layout or SettingsLayout(self, request, _('VAT'))
-    return handle_generic_settings(self, request, form, _('VAT'), layout)
+    layout = layout or SettingsLayout(self, request, _('Value Added Tax'))
+    return handle_generic_settings(self, request, form, _('Value Added Tax'),
+                                   layout)
 
 
 @OrgApp.form(
