+++ src/onegov/town6/views/settings.py
@@ -452,7 +452,7 @@ def town_handle_ticket_data_deletion_settings(
 
 @TownApp.form(
     model=Organisation, name='vat-settings', template='form.pt',
-    permission=Secret, form=VATSettingsForm, setting=_('VAT'),
+    permission=Secret, form=VATSettingsForm, setting=_('Value Added Tax'),
     icon='fa-file-invoice-dollar', order=450)
 def handle_vat_settings(
     self: Organisation,
@@ -460,5 +460,6 @@ def handle_vat_settings(
     form: VATSettingsForm,
     layout: SettingsLayout | None = None
 ) -> RenderData | Response:
-    layout = layout or SettingsLayout(self, request, _('VAT'))
-    return handle_generic_settings(self, request, form, _('VAT'), layout)
+    layout = layout or SettingsLayout(self, request, _('Value Added Tax'))
+    return handle_generic_settings(self, request, form, _('Value Added Tax'),
+                                   layout)
